Выносит константы в utils/constants
