+++ src/components/review/review.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import moment from 'moment';
+import moment from "moment";
 
 const Review = ({quote, author, datetime, rating}) => {
