+++ src/components/tab-details/tab-details.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import moment from 'moment';
+import moment from "moment";
 
 const TabDetails = ({director, starring, runtime, genre, released}) => {
   const formattingList = (list) => {
