+++ src/store/reducers/index.js
@@ -1,7 +1,7 @@
 import {combineReducers} from "redux";
 import {appState} from "./app-state/app-state";
 import {user} from "./user/user";
-import {NameSpace} from "../../utils/utils";
+import {NameSpace} from "../../utils/constants";
 
 export default combineReducers({
   [NameSpace.APP_STATE]: appState,
