+++ src/utils/constants.js
@@ -1,0 +1,23 @@
+export const ALL_GENRE = `All genres`;
+export const FILMS_COUNT_PER_CLICK = 8;
+
+export const isFavoriteType = {
+  CHECKED: true,
+  UNCHECKED: false,
+};
+
+export const TabsType = {
+  OVERVIEW: `Overview`,
+  DETAILS: `Details`,
+  REVIEWS: `Reviews`
+};
+
+export const NameSpace = {
+  APP_STATE: `APP_STATE`,
+  USER: `USER`
+};
+
+export const AuthorizationStatus = {
+  AUTH: `AUTH`,
+  NO_AUTH: `NO_AUTH`,
+};
